Iterator: New `Buffer` iterator.
This iterator acts as a buffer on another iterator. We can go forward as
usual, but we can also go backward over already computed pairs. A buffer
size ∈ [1;+∞] must be defined. When going backward and forward, already
computed pairs are not re-computed, except when forward and when
necessary. When the buffer is full, the oldest computed value is shifted
from it. Thus:

    $abcde  = new Hoa\Iterator\Map(['a', 'b', 'c', 'd', 'e']);
    $buffer = new Hoa\Iterator\Buffer($abcde, 3);

    $buffer->rewind();
    echo $buffer->current(); // a

    $buffer->next();
    echo $buffer->current(); // b

    $buffer->next();
    echo $buffer->current(); // c

    $buffer->previous();
    echo $buffer->current(); // b

    $buffer->previous();
    echo $buffer->current(); // a

    $buffer->next();
    echo $buffer->current(); // b

    /**
     * Will output:
     *     abcbab
     */